Modificacion: base de datos para datos mas grandes costos fijos y guardado de estudio financiero

+++ database/migrations/2024_04_02_224522_create_estudio_financieros_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             // Los campos de la tabla
             $table->id();
             $table->foreignId('plan_de_negocio_id')->unique()->constrained('plan_de_negocios')->onDelete('cascade');
-            $table->float('total_costo_fijo');
-            $table->float('total_costo_variable');
-            $table->float('total_ingresos');
+            $table->decimal('total_costo_fijo',12,2);
+            $table->decimal('total_costo_variable',12,2);
+            $table->decimal('total_ingresos',12,2);
             $table->timestamps();
         });
     }
